+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## Trunk
 
+* router: split writer to stdout and stderr
 * types: validate integer option is not NaN
 * router: route no longer accept process
 * router: remove end property